Add a content Page
Add a new content page to the views Folder and name it WelcomePage. Add a parks page with a list view, on the model page add a new class and name it park, setup Mvvm Helpers and viewModel inheritance. Add a new class named ParksViewModel to the ViewModels. Setup a park service using dependency injection, add a service that will connect the source data. Add a new interface class called IParkDataStore.