ListView, Selection and Events
On ParksPage add resource for the content page, create a dictionary of resources. Add a public variable to the ParksViewModel.cs file. Add events in the ParksPark.xaml, make a select command in the ParksViewModel.cs. Add a park Detail page to the views folder, bind the tile to the name. Register the route of the newly added page else the AppShell will not know the page. In the ParkDetailPage it should be know that it will receive some data.