Added more features to the app
For the parks page, i changed the text color for both Name and description, then the font attributes, font size and also background color. For the ParksDetailPage for the name and description i included font size, font attribute and background color. In the Welcome Page,  i included student name, student number and set Text color, font attributes, font size. Also included the description of the ParksApp, benefit of the app, what makes the app fun for family, feature of the ParksApp, pages and their function and lastly the technical challenges for the ParksApp